thesis: replace auto-fill-mode format
This reverts the paragraph and sentence structure to how I otherwise
would write my thesis directly in LaTex: line breaks are done where one
would pause for breath when reading the text aloud. This is usually at
the end of sentences, but also at some semi-colons.

This structure is easier to write and edit (but also more fun). For
example, it is now very easy to spot mistakes such as starting a chain
of sentencens with "The". Git diffs will also be improved: a word change
in a paragragh now yield a single addition and removal, instead of
multiple where the paragraph would have to be reflowed.