typesetting: breakline most overlong words
This one is a doozy.  The problem: some \texttt strings and words
containing \textunderscore are not automatically breaklined, leading to
overfull hboxes.

The \textunderscore sue is trivial:

        \usepackage[htt]{hyphenat}

and replace all instances of \textunderscore with \_.  These no longer
intrude in the margin, but produce a hypen to denote the break.  Not
ideal, but I'll allow it.  This hyphen only occurs for registers in the
glossary for now.

The other issue is not trivial: \texttt can be redefined alike the added
latex headers in this commit, but then compilation fails \texttt usages
in listing captions and section names.  Alright, so we just create a
\textttBreak that breaks on / : - and _, common characters in long
strings one would \texttt.  But how do we use it?
org-latex-export-to-pdf translates both ~some text~ and =some text= to
\texttt{some text}.  Can we make to that ~some text~ exports to
\textttBreak{some text}?

Some time later after digging through nine year old org-mode mailing
list archives and by happenstance stumbling upon the currently-in-use
org-latex-text-markup-alist variable it turns out that we can!  By
default its value is

        ((bold . "\\textbf{%s}")
          (code . protectedtexttt)
          (italic . "\\emph{%s}")
          (strike-through . "\\sout{%s}")
          (underline . "\\uline{%s}")
          (verbatim . protectedtexttt)))

the rhs value being some kind of symbol, meaning that we cannot
trivially override the usage of \texttt, while also escaping characters
LaTeX will panic on.

The solution? Just copy a subset of not-for-public-use functions from
ox-latex.el in the org-mode source, modify a few a line of two, and live
happily ever after after [0]. Oh, and make sure to use ~some text~
inside captions and section headings.

[0] That is: until org-mode updates and the functions I have overridden
    have been replaced.

Note that this commit does not fix all overfull hboxes. A few seeminly
trivial errors still exist, but I think I can just trim the document
geography to fix those.

LaTeX + Org-mode abstraction is something special.